Gridify sidebar with 3 row divs
Instead of having a 2D grid running down the sidebar, we make a grid
of 3 row divs. Then each div is made a grid with 2 columns and
appropriate number of rows and spacing. This can allow better alignment
of elements on the page. Further, sidebar needs to have a fourth row
to shrink size of each grid cell in the other 3 row divs.