Align titles in feed section correctly
If titles are part of the grid as before, changing row gap also changes
gap between title and first card. So, titles must be taken outside of
cards grid. A row in the common container must be added to place titles,
and the row gap between the two rows should be ~ 0.