Add trending card
It is not possible in the earlier layout to make a title "Trending" as a
new row would have to be added and then row-gap would be used. So,
instead, we create a new div which has a grid inside and contains the
title "Trending" and the corresponding card. We allow the grid to
stretch to the maximum possible.